Some methods moved to DonationsController
