--- src/main/resources/templates/user/user-panel.html.bak
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="pl" xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org"
-      xmlns:sec="https://www.thymeleaf.org/thymeleaf-extras-springsecurity5">
-<head>
-    <meta charset="UTF-8">
-    <title>User-Panel</title>
-</head>
-<body>
-<h2>USER PANEL</h2>
-<div sec:authorize="isAuthenticated()">
-    Authenticated as <span sec:authentication="name"></span></div>
-<h2 th:text="${userName}">NAME</h2>
-<!--<h2 th:text="${userEmail}">NAME</h2>-->
-</body>
-</html>
